+++ Plugins/Purple Service/ESSimpleService.m
@@ -50,7 +50,7 @@
 - (NSCharacterSet *)ignoredCharacters{
 	return [NSCharacterSet characterSetWithCharactersInString:@""];
 }
-- (int)allowedLength{
+- (NSUInteger)allowedLength{
 	return 255;
 }
 - (BOOL)caseSensitive{
