+++ Plugins/Purple Service/PurpleMySpaceService.m
@@ -48,7 +48,7 @@
 - (NSCharacterSet *)ignoredCharacters{
 	return [NSCharacterSet characterSetWithCharactersInString:@""];
 }
-- (int)allowedLength{
+- (NSUInteger)allowedLength{
 	return 999;
 }
 - (BOOL)caseSensitive{
